Project open "drawer" functionality: no longer hiding thumbnail when a project is open
In order to simplify presentation and tricky functionality issues
(i.e., detecting the order of the project in the line of projects and
then using that information to detect which sibling <li> would be the
basis for the height of the open container), switched to an approach
that keeps the thumbnail visible in the line of projects and uses that
as the basis for the open drawer height. This way it doesn't matter
number the project is in the line of projects, as its own <li> will be
used to formulate the height of the drawer.